Install helpful tools for pipeline engineering (moreutils)
moreutils provides a few helpful utilities that combine well with tools like
`kustomize` when you need to preprocess or process in- and output of pipes.

For example, `envsubst < kustomization.yaml | sponge kustomization.yaml` allows in-place
modification of a file w/o race conditions that the `tee` command would be subject to.