Re-install tzdata
Even though `rpm -q tzdata` shows that the tzdata is installed, it is
not actually available. It appears that `/usr/share/zoneinfo` is
stripped from the ubi9-minimal image to save space.

We re-install the tzdata package in order to have time zone information
available, e.g. to parse and format dates with yq.