Fix empty `/usr/share/zoneinfo`
RedHat BUM images contain a zeroed out `/usr/share/zoneinfo` to keep the images (too) lightweight.

The tzdata package is not fully removed because of dependency issues. Because of this an `install` is not enough to trigger the generation of a valid zoneinfo file.

This PR adds and update and reinstall step for tzdata as per https://access.redhat.com/solutions/5616681.